# Rewrite ConcurrentQueue<T> for better performance