--- a/src/System.Collections.Concurrent/src/Resources/Strings.resx
+++ b/src/System.Collections.Concurrent/src/Resources/Strings.resx
@@ -183,7 +183,7 @@
   <data name="ConcurrentBag_CopyTo_ArgumentNullException" xml:space="preserve">
     <value>The array argument is null.</value>
   </data>
-  <data name="ConcurrentBag_CopyTo_ArgumentOutOfRangeException" xml:space="preserve">
+  <data name="Collection_CopyTo_ArgumentOutOfRangeException" xml:space="preserve">
     <value>The index argument must be greater than or equal zero.</value>
   </data>
   <data name="ConcurrentCollection_SyncRoot_NotSupported" xml:space="preserve">