--- a/src/System.Collections.Concurrent/tests/ProducerConsumerCollectionTests.cs
+++ b/src/System.Collections.Concurrent/tests/ProducerConsumerCollectionTests.cs
@@ -2,6 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+// #define StressTest // set to raise the amount of time spent in concurrency tests that stress the collections
+
 using System.Collections.Generic;
 using System.Collections.Tests;
 using System.Diagnostics;
@@ -30,6 +32,12 @@ namespace System.Collections.Concurrent.Tests
 
         protected static TaskFactory ThreadFactory { get; } = new TaskFactory(
             CancellationToken.None, TaskCreationOptions.LongRunning, TaskContinuationOptions.LongRunning, TaskScheduler.Default);
+        private const double ConcurrencyTestSeconds =
+#if StressTest
+            8.0;
+#else
+            1.0;
+#endif
 
         [Fact]
         public void Ctor_InvalidArgs_Throws()
@@ -117,7 +125,7 @@ namespace System.Collections.Concurrent.Tests
             IProducerConsumerCollection<int> c = CreateOracle();
             IProducerConsumerCollection<int> oracle = CreateProducerConsumerCollection();
 
-            Action dequeue = () =>
+            Action take = () =>
             {
                 int item1;
                 Assert.True(c.TryTake(out item1));
@@ -138,14 +146,14 @@ namespace System.Collections.Concurrent.Tests
                 // Start taking some after we've added some
                 if (i > 50)
                 {
-                    dequeue();
+                    take();
                 }
             }
 
             // Take the rest
             while (c.Count > 0)
             {
-                dequeue();
+                take();
             }
         }
 
@@ -250,9 +258,9 @@ namespace System.Collections.Concurrent.Tests
                 Assert.Equal(oracle.TryAdd(initialCount + i), c.TryAdd(initialCount + i));
                 Assert.Equal<int>(oracle, c);
 
-                int expectedDequeued, actualDequeued;
-                Assert.Equal(oracle.TryTake(out expectedDequeued), c.TryTake(out actualDequeued));
-                Assert.Equal(expectedDequeued, actualDequeued);
+                int expected, actual;
+                Assert.Equal(oracle.TryTake(out expected), c.TryTake(out actual));
+                Assert.Equal(expected, actual);
                 Assert.Equal<int>(oracle, c);
             }
         }
@@ -264,7 +272,7 @@ namespace System.Collections.Concurrent.Tests
 
             for (int i = 0; i < 1000; i += 100)
             {
-                // Create a thread that adds items to the bag
+                // Create a thread that adds items to the collection
                 ThreadFactory.StartNew(() =>
                 {
                     for (int j = i; j < i + 100; j++)
@@ -297,7 +305,7 @@ namespace System.Collections.Concurrent.Tests
         }
 
         [Fact]
-        public void TryPeek_SucceedsOnEmptyBagThatWasOnceNonEmpty()
+        public void TryPeek_SucceedsOnEmptyCollectionThatWasOnceNonEmpty()
         {
             IProducerConsumerCollection<int> c = CreateProducerConsumerCollection();
             int item;
@@ -340,7 +348,7 @@ namespace System.Collections.Concurrent.Tests
                 while (!cts.IsCancellationRequested) Assert.False(IsEmpty(c));
             });
 
-            // Producer enqueues/dequeues a bunch of items, then tells the consumer to stop
+            // Producer adds/takes a bunch of items, then tells the consumer to stop
             Task producer = Task.Run(() =>
             {
                 int ignored;
@@ -421,6 +429,26 @@ namespace System.Collections.Concurrent.Tests
         }
 
         [Fact]
+        public void CopyTo_ArrayNonZeroLowerBound_ExpectedElementsCopied()
+        {
+            int[] initialItems = Enumerable.Range(1, 10).ToArray();
+
+            const int LowerBound = 1;
+            ICollection c = CreateProducerConsumerCollection(initialItems);
+            Array actual = Array.CreateInstance(typeof(int), new int[] { initialItems.Length }, new int[] { LowerBound });
+            c.CopyTo(actual, LowerBound);
+
+            ICollection oracle = CreateOracle(initialItems);
+            int[] expected = new int[initialItems.Length];
+            oracle.CopyTo(expected, 0);
+
+            for (int i = 0; i < expected.Length; i++)
+            {
+                Assert.Equal(expected[i], actual.GetValue(i + LowerBound));
+            }
+        }
+
+        [Fact]
         public void CopyTo_InvalidArgs_Throws()
         {
             IProducerConsumerCollection<int> c = CreateProducerConsumerCollection(Enumerable.Range(0, 10));
@@ -430,24 +458,26 @@ namespace System.Collections.Concurrent.Tests
             Assert.Throws<ArgumentOutOfRangeException>(() => c.CopyTo(dest, -1));
             Assert.Throws<ArgumentException>(() => c.CopyTo(dest, dest.Length));
             Assert.Throws<ArgumentException>(() => c.CopyTo(dest, dest.Length - 2));
+
+            Assert.Throws<ArgumentException>(() => c.CopyTo(new int[7, 7], 0));
         }
 
         [Fact]
         public void ICollectionCopyTo_InvalidArgs_Throws()
         {
-            ICollection bag = CreateProducerConsumerCollection(Enumerable.Range(0, 10));
+            ICollection c = CreateProducerConsumerCollection(Enumerable.Range(0, 10));
             Array dest = new int[10];
 
-            Assert.Throws<ArgumentNullException>("array", () => bag.CopyTo(null, 0));
-            Assert.Throws<ArgumentOutOfRangeException>(() => bag.CopyTo(dest, -1));
-            Assert.Throws<ArgumentException>(() => bag.CopyTo(dest, dest.Length));
-            Assert.Throws<ArgumentException>(() => bag.CopyTo(dest, dest.Length - 2));
+            Assert.Throws<ArgumentNullException>("array", () => c.CopyTo(null, 0));
+            Assert.Throws<ArgumentOutOfRangeException>(() => c.CopyTo(dest, -1));
+            Assert.Throws<ArgumentException>(() => c.CopyTo(dest, dest.Length));
+            Assert.Throws<ArgumentException>(() => c.CopyTo(dest, dest.Length - 2));
         }
 
         [Theory]
         [InlineData(100, 1, 10)]
         [InlineData(4, 100000, 10)]
-        public void BlockingCollection_WrappingBag_ExpectedElementsTransferred(int numThreadsPerConsumerProducer, int numItemsPerThread, int producerSpin)
+        public void BlockingCollection_WrappingCollection_ExpectedElementsTransferred(int numThreadsPerConsumerProducer, int numItemsPerThread, int producerSpin)
         {
             var bc = new BlockingCollection<int>(CreateProducerConsumerCollection());
             long dummy = 0;
@@ -662,13 +692,14 @@ namespace System.Collections.Concurrent.Tests
         }
 
         [Theory]
-        [InlineData(1.0)]
-        public void ManyConcurrentAddsTakes_OneThreadPerProc(double seconds)
+        [InlineData(1, ConcurrencyTestSeconds / 2)]
+        [InlineData(4, ConcurrencyTestSeconds / 2)]
+        public void ManyConcurrentAddsTakes_EnsureTrackedCountsMatchResultingCollection(int threadsPerProc, double seconds)
         {
             IProducerConsumerCollection<int> c = CreateProducerConsumerCollection();
 
             DateTime end = default(DateTime);
-            using (var b = new Barrier(Environment.ProcessorCount, _ => end = DateTime.UtcNow + TimeSpan.FromSeconds(seconds)))
+            using (var b = new Barrier(Environment.ProcessorCount * threadsPerProc, _ => end = DateTime.UtcNow + TimeSpan.FromSeconds(seconds)))
             {
                 Task<int>[] tasks = Enumerable.Range(0, b.ParticipantCount).Select(_ => ThreadFactory.StartNew(() =>
                 {
@@ -699,7 +730,7 @@ namespace System.Collections.Concurrent.Tests
         }
 
         [Theory]
-        [InlineData(1.0)]
+        [InlineData(ConcurrencyTestSeconds)]
         public void ManyConcurrentAddsTakes_CollectionRemainsConsistent(double seconds)
         {
             IProducerConsumerCollection<int> c = CreateProducerConsumerCollection();
@@ -769,12 +800,12 @@ namespace System.Collections.Concurrent.Tests
             total.ExceptWith(addsAndTakes.Result.Value);
             total.ExceptWith(takes.Result);
 
-            // What's left should match what's in the bag
+            // What's left should match what's in the collection
             Assert.Equal(total.OrderBy(i => i), c.OrderBy(i => i));
         }
 
         [Theory]
-        [InlineData(1.0)]
+        [InlineData(ConcurrencyTestSeconds)]
         public void ManyConcurrentAddsTakesPeeks_ForceContentionWithOtherThreadsTaking(double seconds)
         {
             IProducerConsumerCollection<int> c = CreateProducerConsumerCollection();
@@ -833,7 +864,7 @@ namespace System.Collections.Concurrent.Tests
         }
 
         [Theory]
-        [InlineData(1.0)]
+        [InlineData(ConcurrencyTestSeconds)]
         public void ManyConcurrentAddsTakesPeeks_ForceContentionWithOtherThreadsPeeking(double seconds)
         {
             IProducerConsumerCollection<int> c = CreateProducerConsumerCollection();
@@ -886,7 +917,7 @@ namespace System.Collections.Concurrent.Tests
         }
 
         [Theory]
-        [InlineData(1.0)]
+        [InlineData(ConcurrencyTestSeconds)]
         public void ManyConcurrentAddsTakes_ForceContentionWithToArray(double seconds)
         {
             IProducerConsumerCollection<int> c = CreateProducerConsumerCollection();
@@ -923,8 +954,8 @@ namespace System.Collections.Concurrent.Tests
         }
 
         [Theory]
-        [InlineData(0, 0.5)]
-        [InlineData(1, 0.5)]
+        [InlineData(0, ConcurrencyTestSeconds / 2)]
+        [InlineData(1, ConcurrencyTestSeconds / 2)]
         public void ManyConcurrentAddsTakes_ForceContentionWithGetEnumerator(int initialCount, double seconds)
         {
             IProducerConsumerCollection<int> c = CreateProducerConsumerCollection(Enumerable.Range(1, initialCount));