@i3arnon 2016-12-06
@stephentoub  these improvements to the concurrent collections are awesome. Will they also be ported back to the full framework? Or are they limited to corefx?