@benaadams 2016-12-07
@i3arnon I assume you can just reference the Corefx versions from full framework using nuget? https://www.nuget.org/packages/System.Collections.Concurrent/ (might be wrong...)