@i3arnon 2016-12-07
@benaadams there's no actual dll for net45 in that package. I assume it just takes the one in the installed FX (or the GAC?).  Not sure though... assembly loading & precedence was never my strong suit.
![image](https://cloud.githubusercontent.com/assets/6595156/20949476/26a6d3b6-bc22-11e6-9196-de54a5688fe2.png)