@stephentoub 2016-12-07
@i3arnon, thanks.  It's not automatically ported back to desktop, but I've labeled it as netfx-port-consider, which has the effect of opening an internal work item to consider doing so.