@benaadams 2016-12-07
Awesome change! Does this also need to be ported to coreclr [ConcurrentQueue](https://github.com/dotnet/coreclr/blob/master/src/mscorlib/src/System/Collections/Concurrent/ConcurrentQueue.cs)? Not sure how the two types interact?

Aside: Comparing the results with ConcurrentBag https://github.com/dotnet/corefx/pull/14126 does this still make ConcurrentQueue the preferred go to data structure of the two? (and more so after this change)

Are there any learnings that could be applied back to ConcurrentBag; as in theory since its is weaker ordered and makes use of thread locals could be faster due to less contention. Or is it the differential because the example perf tests are always hitting an enumeration of steal; which is worse case for Bag, than Queue - so is more the scenario tested?