@stephentoub 2016-12-07
> Does this also need to be ported to coreclr ConcurrentQueue? Not sure how the two types interact?

The ConcurrentQueue that's internal to coreclr has minimal usage.  I don't think it's worthwhile for performance reasons; the primary benefit would be in keeping them in sync.

> Comparing the results with ConcurrentBag #14126 does this still make ConcurrentQueue the preferred go to data structure of the two? (and more so after this change)

"go to" for what purpose?  First, the results I included in the two PRs aren't directly comparable.  And which data structure makes more sense will depend on the situation.  For example, if you're doing something producer/consumer, with dedicated producers and dedicated consumers, ConcurrentQueue is going to be better than ConcurrentBag, e.g. this test:
```C#
using System;
using System.Collections;
using System.Collections.Concurrent;
using System.Diagnostics;
using System.Threading.Tasks;

class Program
{
    static void Main()
    {
        const int Iters = 10000000;
        var sw = new Stopwatch();
        while (true)
        {
            var cb = new ConcurrentBag<int>();
            var cq = new ConcurrentQueue<int>();

            sw.Restart();
            Task.WaitAll(
                Task.Run(() =>
                {
                    for (int i = 0; i < Iters; i++) cb.Add(i);
                }),
                Task.Run(() =>
                {
                    int item;
                    for (int i = 0; i < Iters; i++) while(!cb.TryTake(out item));
                }));
            sw.Stop();
            Console.WriteLine("Bag  : " + sw.Elapsed);
            GC.Collect();

            sw.Restart();
            Task.WaitAll(
                Task.Run(() =>
                {
                    for (int i = 0; i < Iters; i++) cq.Enqueue(i);
                }),
                Task.Run(() =>
                {
                    int item;
                    for (int i = 0; i < Iters; i++) while(!cq.TryDequeue(out item));
                }));

            sw.Stop();
            Console.WriteLine("Queue: " + sw.Elapsed);
            GC.Collect();

            Console.WriteLine();
        }
    }
}
```
on my machine has ConcurrentQueue performing 5x better than ConcurrentBag.  But for example a situation where each thread is doing its own adds and takes but wants to have the whole collection available to it for takes if needed, ConcurrentBag scales much better, e.g. this test:
```C#
using System;
using System.Collections;
using System.Collections.Concurrent;
using System.Diagnostics;
using System.Threading.Tasks;

class Program
{
    static void Main()
    {
        const int Iters = 10000000;
        var sw = new Stopwatch();
        while (true)
        {
            var cb = new ConcurrentBag<int>();
            var cq = new ConcurrentQueue<int>();

            sw.Restart();
            Parallel.For(0, Iters, i =>
            {
                int item;
                cb.Add(i);
                cb.TryTake(out item);
            });
            sw.Stop();
            Console.WriteLine("Bag  : " + sw.Elapsed);
            GC.Collect();

            sw.Restart();
            Parallel.For(0, Iters, i =>
            {
                int item;
                cq.Enqueue(i);
                cq.TryDequeue(out item);
            });
            sw.Stop();
            Console.WriteLine("Queue: " + sw.Elapsed);
            GC.Collect();

            Console.WriteLine();
        }
    }
}
```
on my machine has ConcurrentBag performing 5x better than ConcurrentQueue.