@benaadams 2016-12-07
> The ConcurrentQueue that's internal to coreclr has minimal usage.

Wasn't sure if it was forwarded - never know these days 😉 

>  depend on the situation... ConcurrentQueue performing 5x better than ConcurrentBag... ConcurrentBag performing 5x better than ConcurrentQueue.

K, hoping that was so - thank you for verifying.