@benaadams 2016-12-08
LGTM - The approach taken makes sense; and it seems to be doing the right thing; however I've not extensively verified line by line.