@benaadams 2016-12-08
Perhaps add something to the `<remarks>` comment of `TryPeek` that it is a more costly operation, and suggest using `IsEmpty` in preference if its just to determine if there is potentially a next item?