@benaadams 2016-12-08
*Thinking aloud...*

I worry about increasing the cost of Count as it has a [high usage](https://apisof.net/catalog/System.Collections.Concurrent.ConcurrentQueue%3CT%3E.Count) relative to [Enqueue](https://apisof.net/catalog/System.Collections.Concurrent.ConcurrentQueue%3CT%3E.Enqueue(T)) 

Some usage may be replaceable with `IsEmpty`; speculating, another use may be to apply backpressure? 

However, Introducing a "cheaper" way of counting, for example `Interlocked.Increment`/`Decrement` to approximate count would increase the cost for regular operations; when it may not be used which I don't think would be good.

If it is used for backpressure then a wrapping type would need to be used to apply the backpressure; and that could also perform the counting and bear the cost - and there are other bounded Producer/Consumer data structures that already provide this.

So... wondering if a blog post or release note announcement item should go with this change? Just for awareness.