@benaadams 2016-12-08
*Note*: As a follow up, I don't think the performance should be generally compromised for `Count` as there are a lot of use cases that don't use `Count`.

At high throughput it is a bit of a Schrödinger's value; and already different after measuring. It's either high velocity of change/thoughput; point in time accuracy; or the user has a race condition...